rename: 파일들의 위치를 이동
>> 역할에 따라 파일들을 분류하기 위해 파일들의 위치를 이동
