+++ src/main/java/racingcar/view/constant/Message.java
@@ -1,4 +1,4 @@
-package racingcar.view.message;
+package racingcar.view.constant;
 
 public interface Message {
     public String getMessage();
