+++ src/main/java/racingcar/view/constant/PromptMessage.java
@@ -1,4 +1,4 @@
-package racingcar.view.message;
+package racingcar.view.constant;
 
 public enum PromptMessage implements Message {
     CAR_NAME_INPUT("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
